# internal/server/web/openapi.yaml
...
openapi: 3.1.0
info:
  title: wudict
  version: "1"
  summary: A local dictionary server for MDX/MDD, StarDict, Slob, DSL and BGL files.
  description: |
    WuWeiDict serves your own dictionary files over HTTP, on your own machine.

    **Streaming.** `/api/dicts`, `/api/search` and `/api/rescan` answer with
    NDJSON: one JSON object per line, flushed the moment it is ready. Read them
    line by line; do not wait for the body to end. That is the whole point - the
    first dictionary's results arrive while the slowest one is still opening.
    Every stream opens with a `begin` line and closes with an `end` line.

    `/api/ingest` streams Server-Sent Events instead, because it reports
    progress rather than results.

    **Switches.** Every boolean query parameter reads the same spellings: off
    for `0`, `false`, `no`, `off` or an empty value, on for anything else.

    **Two surfaces.** Three read-only endpoints answer a browser extension
    cross-origin: `/api/dicts`, `/api/search` and `/res/`. Everything else is
    same-origin, is marked `internal` below, and may change between releases.

    **An access key, when the server is reachable.** On the default
    `127.0.0.1` bind there is nothing to authenticate and no key is asked for.
    On any other address - and Go's wildcard listen is dual-stack, so `0.0.0.0`
    includes every globally routable IPv6 address the machine holds - `AUTH`
...
    and let the browser keep the cookie it sets. Without it: `401`, with
...

...

...
servers:
...
security: []
...
paths:
  /api/search:
    get:
      tags: [search]
      operationId: search
      summary: Search every dictionary, streaming
      description: |
        Queries the selected dictionaries concurrently and streams one line per
        dictionary as it finishes. Lines arrive in completion order, never in
        slot order: each `hit` carries `i`, its index in the `begin` line's
        `slots` array.

        A dictionary that fails, that cannot serve the requested mode, or that
        was declined by the memory cap reports that on its own line. None of
        those is a request failure, and the other dictionaries still answer.

        A search is cancelled after 30 seconds.
      parameters:
        - name: q
          in: query
          required: true
          description: The search text. Blank is a 400.
          schema: { type: string }
          example: flight
        - name: mode
          in: query
          description: |
            `exact` · `prefix` · `contains` (trigram substring over folded
            headwords) · `fts` (full text). `fuzzy` is accepted as a legacy
            alias and now means `prefix`.

            `contains` and `fts` need the dictionary to carry that index; ask
            `/api/dicts` for `caps` first, and expect `skipped` if you do not.

            A multi-word `fts` query is a PHRASE first. It is retried as a
            proximity and then as separate words only when the stricter reading
            returns nothing, and the hit says which reading answered (`rung`).
            Quotes and the operators `AND`, `OR`, `NOT` and `NEAR(a b, n)` are
            parsed when present, and an explicit query like
            `"faux ami" NOT linguistics` is answered exactly as written, without
            relaxation. Anything unparseable is read as plain words rather than
            rejected.
          schema:
            type: string
            enum: [exact, prefix, contains, fts, fuzzy]
            default: prefix
        - name: hl
          in: query
          description: |
            Mark the query's words inside each article body, wrapped in
            `<mark class="wu-hl">`. Read ONLY in `mode=fts`, where the reason a
            result matched is invisible otherwise; ignored in every other mode,
            where the match is the headword itself.

            Off unless you ask, so an article you fetch is the dictionary's own
            bytes by default. wudict's own UI sends `hl=1` from a remembered
            preference, which is why the default here is off and the default
            there is on. Marking is prefix-based and whole-token, matching what
            the full-text index matched, and is capped at 500 marks per article
            - past the cap the article is still returned complete, just no
            longer marked. `format=text` drops the marks with all other markup.
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          description: |
            `raw` is the dictionary's own HTML. `clean` keeps structure,
            emphasis and media and drops scripts, styles and presentation, at
            roughly half the bytes; it also rewrites root-absolute links such
            as `/res/…` into absolute URLs, so the payload works inside a page
            served from somewhere else. `text` is markup-free, roughly 0.4.

            Ask for `clean` unless you render the article with the dictionary's
            own stylesheet.
          schema:
            type: string
            enum: [raw, clean, text]
            default: raw
        - name: dict
          in: query
          description: >
            `all`, one dictionary id, or a comma-separated list of ids in the
            order you want them back. Ids come from /api/dicts. A list where no
            id resolves is a 404; unknown ids inside a resolving list are
            skipped. Naming exactly one dictionary bypasses the memory cap,
            because a search of one is a direct demand.
          schema: { type: string, default: all }
          example: oxford,webster
        - name: 'n'
          in: query
          description: Maximum results per dictionary.
          schema: { type: integer, minimum: 1, default: 20 }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -sN 'http://127.0.0.1:6888/api/search?q=flight&mode=prefix&format=clean&n=5'

            # -N is the whole point: without it curl buffers, and the lines that
            # were sent one at a time arrive together at the end.
      responses:
        '200':
          description: The NDJSON stream. One object per line.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SearchBegin'
                  - $ref: '#/components/schemas/SearchHit'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
              example: |
                {"t":"begin","i":0,"slots":[{"dict":"oxford","name":"oxford"},{"dict":"webster","name":"webster"}]}
                {"t":"hit","i":1,"dict":"webster","name":"Webster's Revised Unabridged","results":[{"Headword":"flight","Body":"<div>…</div>"}]}
                {"t":"hit","i":0,"dict":"oxford","name":"Oxford Advanced Learner's","results":[]}
                {"t":"end","i":0}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: searchPreflight
      summary: CORS preflight
      description: >
        Answers 204 with the extension grant. The extension's own requests are
        CORS-simple and never preflight; this exists so a future one does not
        meet a 405 that explains nothing.
      responses:
        '204': { description: Preflight accepted. }
  /api/dicts:
    get:
      tags: [search]
      operationId: listDicts
      summary: List every dictionary, streaming
      description: >
        `begin` carries the total immediately, from registry ids alone, so a
        client can show "0 of 105" before any dictionary is opened. Rows follow
        in completion order.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -sN 'http://127.0.0.1:6888/api/dicts'

            # just the ids, which is what /api/search wants
            curl -sN 'http://127.0.0.1:6888/api/dicts' | jq -r 'select(.t=="dict") | .dict.id'
      responses:
        '200':
          description: The NDJSON stream. One object per line.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DictsBegin'
                  - $ref: '#/components/schemas/DictsRow'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
              example: |
                {"t":"begin","total":2}
                {"t":"dict","dict":{"id":"oxford","name":"Oxford Advanced Learner's","format":"mdx","path":"/Users/me/Dicts/oald.mdx","entries":184000,"caps":{"Exact":true,"Prefix":true,"Contains":false,"FTS":false}}}
                {"t":"end"}
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: dictsPreflight
      summary: CORS preflight
      responses:
        '204': { description: Preflight accepted. }
...
  /api/lemmas:
    get:
      tags: [internal]
      operationId: listLemmas
      x-internal: true
      summary: List installable and installed lemma data
      description: >
        One row per language, merging the published catalogue (LEMMA_URL) with
        what is present in LEMMA_DIR, plus the state of any download running
        right now. Always answers 200: what is installed is the half of the
        answer that needs no network, and an unreachable catalogue is reported
        in `error` rather than by failing.
      parameters:
        - name: refresh
          in: query
          description: >
            Re-read the catalogue instead of using the cached copy. The
            catalogue is otherwise cached for ten minutes, so a page polling
            during a download does not ask its host once a second.
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s 'http://127.0.0.1:6888/api/lemmas' |
              jq -r '.languages[] | select(.installed) | .code'
      responses:
        '200':
          description: The catalogue merged with the folder.
          content:
            application/json:
              schema:
                type: object
                properties:
                  dir: { type: string, description: LEMMA_DIR. }
                  url: { type: string, description: The catalogue this lists. }
                  enabled:
                    type: boolean
                    description: >
                      False when MORPH_CACHE is 0, in which case nothing
                      installed here will ever be consulted by a search.
                  cacheSize:
                    type: integer
                    description: How many language packs may be resident at once.
                  attribution:
                    type: string
                    description: ATTRIBUTION.txt published beside the catalogue (ODbL).
                  error:
                    type: string
                    description: Why the catalogue could not be read, if it could not.
                  shadowed:
                    type: array
                    description: Files the lemmatizer ignores because another supplies the same language.
                    items: { type: string }
                  languages:
                    type: array
                    items: { $ref: '#/components/schemas/LemmaLanguage' }
    post:
      tags: [internal]
      operationId: installLemmas
      x-internal: true
      summary: Install one language's lemma data
      description: |
        Starts a download and returns at once; follow it with `GET /api/lemmas`.
        The download runs in the server, not in this request, so it survives the
        page being reloaded, closed, or rotated away on Android.

        The catalogue is never taken from the caller: only the configured
        LEMMA_URL is fetched. A second call for a language already downloading
        joins the running download instead of starting another.

        On success the language is searchable immediately - no restart, because
        the lemma folder is re-indexed in place.
      parameters:
        - name: code
          in: query
          required: true
          description: Language code or English name ("pl", "polish").
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s -X POST 'http://127.0.0.1:6888/api/lemmas?code=pl'
      responses:
        '202':
          description: The download is running.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  state: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '502':
          description: The catalogue could not be read, so nothing can be installed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [internal]
      operationId: removeLemmas
      x-internal: true
      summary: Delete one language's lemma data
      description: >
        Removes every file in LEMMA_DIR supplying this language, not only the
        one in use, and re-indexes the folder. Deleting English falls back to
        the pack compiled into the binary rather than losing the language.
        Unlike dictionary removal this is offered on every platform: these are
        files wudict downloaded, not files the user brought.
      parameters:
        - name: code
          in: query
          required: true
          description: Language code or English name.
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s -X DELETE 'http://127.0.0.1:6888/api/lemmas?code=pl'
      responses:
        '200':
          description: What was removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  builtin: { type: boolean }
                  removed:
                    type: array
                    items: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: That language is being downloaded right now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/setup:
    get:
      tags: [internal]
      operationId: setup
      x-internal: true
      summary: Validate a dictionary folder, and optionally save it
      description: >
        Drives the first-run page. Without `save` it only reports what the
        folder holds, which is what live typing needs. With `save=1` it switches
        the registry over and writes DICT_DIR. Repeat `path` for several
        folders; `found` is then the deduplicated total across all of them,
        because overlapping folders must not be counted twice.
      parameters:
        - name: path
          in: query
          description: A dictionary folder. Repeatable. `~` is expanded.
          schema: { type: string }
        - name: save
          in: query
          description: On persists the folders to the config file. A switch.
          schema: { type: string }
        - name: useCached
          in: query
          description: >
            On enrolls the previously prepared dictionaries and persists
            USE_CACHED; off turns that off; absent leaves it alone. A switch.
          schema: { type: string, enum: ['0', '1'] }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # what this folder holds, without changing anything
            curl -s 'http://127.0.0.1:6888/api/setup?path=~/Dictionaries'

            # adopt two folders and write them to the config file
            curl -s 'http://127.0.0.1:6888/api/setup?path=~/Dictionaries&path=/Volumes/Ext/Dicts&save=1'
      responses:
        '200':
          description: >
            The result. A folder that does not exist, holds nothing, or cannot
            be read reports that in `error` with a 200: this answers a keystroke,
            and a half-typed path is not a failed request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/groups:
    get:
      tags: [internal]
      operationId: getGroups
      x-internal: true
      summary: The groups the dictionary picker offers
      description: >
        `groups.ini`, beside the `wudict.toml` in effect: a `[Section]` line is
        a "Group by" heading (at most 32, names of at most 64 bytes), and
        `Name = text, text` a group under it that a dictionary joins when one
        of the texts (at least 2 characters) is anywhere in its title or file
        name, in any case. An
        item between backticks is a regular expression (Go RE2,
        case-insensitive), read verbatim and tested against the title and the
        file name separately. `;` and `#` start a comment outside backticks. A
        group line before any section belongs to `[My groups]`. While there is no file the built-in list is in effect
        (`custom` false); once saved, the file is the whole list.


        Ids are the names, trimmed and lower-cased: a dictionary row's
        `groups[].f` and `groups[].v` (`GET /api/dicts`) are exactly these.
        `problems` lists the lines and items that were skipped (an invalid
        regular expression, a line with no `=`, a one-character item, more
        than 500 regular expressions); the rest of the file still applies. At
        most 100 are listed: a last entry with `more` stands for the rest.
        `text` and `msg` are clipped to 200 bytes. `unusable` means the file exists but could not be read: the
        built-in list is in effect, and `problems` has one entry, with no line,
        saying why.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s 'http://127.0.0.1:6888/api/groups' | jq -r .text
      responses:
        '200':
          description: The file in effect.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Groups' }
    put:
      tags: [internal]
      operationId: saveGroups
      x-internal: true
      summary: Replace groups.ini
      description: >
        The body is the whole file, as text. Lines that do not parse are
        reported in `problems`, never refused. A body identical to the built-in
        list removes the file instead, so an unchanged copy does not stop the
        list from following the built-in one. Capped at 256 KiB. While the
        current file is `unusable`, a save is refused unless `replace=1` is
        given: the file is the user's, and the editor could not show it.
      parameters:
        - name: replace
          in: query
          required: false
          description: 1 overwrites a groups.ini that exists but cannot be read.
          schema: { type: string, enum: ['1'] }
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
      responses:
        '200':
          description: The file in effect after the save.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Groups' }
        '400':
          description: The body is not UTF-8 text, or could not be read.
        '409':
          description: >
            There is no config directory, so there is nowhere to save; or the
            current file cannot be read and `replace=1` was not given.
        '413':
          description: The body is larger than 256 KiB.
        '500':
          description: The file could not be written.
    delete:
      tags: [internal]
      operationId: resetGroups
      x-internal: true
      summary: Back to the built-in groups
      description: Removes groups.ini; the built-in list is in effect again.
      responses:
        '200':
          description: The built-in list, now in effect.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Groups' }
        '500':
          description: The file could not be removed.
...
components:
  responses:
    BadRequest:
      description: A parameter is missing or unknown.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such dictionary id, or no such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: The dictionary could not be opened, or the stream could not start.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Every failing request answers with this shape.
      properties:
        error: { type: string }
      required: [error]
...
    LemmaLanguage:
      type: object
      description: >
        One language, merging the catalogue entry with what is on disk and with
        any download running right now. A language missing from the catalogue
        but present in the folder still appears - it works, so it is listed as
        working.
      properties:
        code: { type: string }
        name: { type: string }
        size:
          type: integer
          description: Compressed download size from the catalogue, in bytes.
        localSize:
          type: integer
          description: Size of the installed file, in bytes. Zero when not installed.
        lemmas: { type: integer, description: Entry count from the catalogue. }
        heapMB:
          type: integer
          description: Measured resident cost once this pack is loaded.
        source: { type: string }
        license: { type: string }
        installed: { type: boolean }
        catalogued:
          type: boolean
          description: False for a file present in the folder but absent from the catalogue.
        builtin:
          type: boolean
          description: True for English, which is compiled in and cannot be removed to nothing.
        mismatch:
          type: boolean
          description: Installed, but its digest is not the catalogue's.
        state:
          type: string
          description: '"downloading" or "error" while a job exists; empty otherwise.'
        done: { type: integer }
        total: { type: integer }
        error: { type: string }
...
    Result:
      type: object
      description: One article. Field names are capitalised on the wire.
      properties:
        Headword: { type: string }
        Body:
          type: string
          description: Article HTML, in the requested format.
    StreamEnd:
      type: object
      description: The last line of any NDJSON stream.
      properties:
        t: { const: end }
        i:
          type: integer
          description: Always 0 here. The search stream carries the field on every line.
      required: [t]
    SearchBegin:
      type: object
      description: >
        The slot layout, sent before any dictionary is opened. Draw the empty
        result areas from it at once; each real name arrives with its hit.
      properties:
        t: { const: begin }
        i: { type: integer, description: Always 0 here. }
        slots:
          type: array
          items:
            type: object
            properties:
              dict: { type: string, description: Dictionary id. }
              name: { type: string, description: The id again; the display name arrives with the hit. }
      required: [t, slots]
    SearchHit:
      type: object
      description: One dictionary's outcome. Fill slot `i`.
      properties:
        t: { const: hit }
        i: { type: integer, description: Index into the begin line's slots array. }
        dict: { type: string }
        name: { type: string, description: The dictionary's real display name. }
        results:
          type: array
          items: { $ref: '#/components/schemas/Result' }
        skipped:
          type: boolean
          description: This dictionary does not support the requested mode.
        deferred:
          type: boolean
          description: >
            Not opened, because the search reached its memory cap. Not an error:
            asking for this dictionary alone answers it. See SEARCH_MEMORY.
        bytes:
          type: integer
          description: What the declined open was estimated to cost. For logs.
        indexing:
          type: boolean
          description: >
            Only beside deferred: this dictionary is already being prepared,
            because something asked for it by name. It will stop being deferred
            once the ingest finishes.
        rung:
          type: string
          enum: [phrase, near, words, query]
          description: >
            Which reading of a full-text query this dictionary answered.
            `phrase` = the words adjacent and in order; `near` = within ten
            tokens of one another, any order; `words` = anywhere in the article;
            `query` = the operator expression as written. A multi-word `fts`
            search tries phrase, then near, then words, and stops at the first
            that returns anything - per dictionary, so two dictionaries in one
            response may answer on different rungs. Absent outside `fts`. Show
            it: a result on `near` or `words` means the phrase itself was not
            found.
        error:
          type: string
          description: This dictionary failed. The others still answered.
      required: [t, i]
    DictsBegin:
      type: object
      properties:
        t: { const: begin }
        total: { type: integer, description: How many rows follow. }
        groupsProblems:
          type: integer
          description: >
            How many lines or items groups.ini skipped (`GET /api/groups` lists
            them); absent when none.
      required: [t]
    DictsRow:
      type: object
      properties:
        t: { const: dict }
        dict: { $ref: '#/components/schemas/DictInfo' }
      required: [t, dict]
...
    Root:
      type: object
      description: One dictionary folder and what it contributed.
      properties:
        path: { type: string }
        count: { type: integer, description: Dictionaries this folder was the first to offer. }
        total: { type: integer, description: Everything it holds. Greater than count means overlap. }
        exists: { type: boolean }
...
    SetupResult:
      type: object
      properties:
        path: { type: string, description: 'The folder, expanded and made absolute.' }
        found: { type: integer, description: 'Dictionaries found, deduplicated across folders.' }
        dirs:
          type: array
          items: { $ref: '#/components/schemas/Root' }
        saved: { type: boolean }
        useCached: { type: boolean }
        error: { type: string, description: Why this folder cannot be used. }